Synths - add support for a wavetable synth (WIP)
This adds a basic synthdef, :wave, that can play a single wavetable file
as an oscillator. Since the path to the wavetable file needs to be
known, the supporting code is similar to the existing code used for
loading samples.

Example syntax for the end user, to play the synth:

synth :wave, wave: ['path/to/wavetables', 'a_wavetable.wavetable']

The array passed in as the value of the `wave:` opt is made up of the
same filter values that are given to the `sample` fn. It's not as neat
as its use in `sample` since it explicitly requires array brackets in
this case, being the value of the `wave` key in the synth args hash.

More work would need to be done to finish this if it is a suitable idea.