Runtime - enable random triggering of 'tagged' snippets
This feature adds the concept of 'tags' to the collection of Sonic Pi code
snippets that a user may decide to collect/make. An extra line can be added to
a snippet file to specify 'tags' with which it can be categorised.
This then allows the user to trigger one of these snippets at random by running a
kind of global 'dummy snippet' and passing it words that match the desired tags
that the random snippet is tagged with. For example:
Say we have some snippets tagged with the words 'beats' and 'broken', where
some are tagged with one of these words, and others with both. We could trigger the
below 'dummy snippet' with the following tag names:

`idk beats, broken`

This would immediately replace itself with the contents of a random snippet
whose tags contained *at least* all of the provided tag names. So, in this case,
it would load the contents of a random snippet that was tagged with *at least*
both 'beats' and 'broken' (in any order).