Scales - allow degrees < 0
This change allows asking for scales with degrees that are less than zero.
However, below zero, the 'degrees' are actually _offsets from the root note_,
rather than absolute degrees. For example, for a scale of C4 major, degree -1
refers to B3, the note in the scale _below degree 1_. Degree 0 is still invalid.